Move example to private and add warning
