+++ packages/react/src/TreeView/TreeView.features.stories.tsx
@@ -4,10 +4,7 @@ import {
   DiffRemovedIcon,
   DiffRenamedIcon,
   FileIcon,
-  GrabberIcon,
   KebabHorizontalIcon,
-  IssueClosedIcon,
-  IssueOpenedIcon,
 } from '@primer/octicons-react'
 import type {Meta, StoryFn} from '@storybook/react'
 import React from 'react'
@@ -996,54 +993,6 @@ export const WithoutIndentation: StoryFn = () => (
     </TreeView>
   </nav>
 )
-
-export const LeadingAction: StoryFn = () => {
-  return (
-    <TreeView aria-label="Issues">
-      <TreeView.Item id="item-0">
-        <TreeView.LeadingAction>
-          <IconButton icon={GrabberIcon} aria-label="Reorder item 1" variant="invisible" />
-        </TreeView.LeadingAction>
-        <TreeView.LeadingVisual label="Issue closed">
-          <Octicon icon={IssueClosedIcon} sx={{color: 'done.fg'}} />
-        </TreeView.LeadingVisual>
-        Item 1
-      </TreeView.Item>
-      <TreeView.Item id="item-2">
-        <TreeView.LeadingAction>
-          <IconButton icon={GrabberIcon} aria-label="Reorder item 2" variant="invisible" />
-        </TreeView.LeadingAction>
-        <TreeView.LeadingVisual label="Issue opened">
-          <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
-        </TreeView.LeadingVisual>
-        Item 2
-        <TreeView.SubTree>
-          <TreeView.Item id="item-2-sub-task-1">
-            <TreeView.LeadingVisual label="Issue opened">
-              <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
-            </TreeView.LeadingVisual>
-            sub task 1
-          </TreeView.Item>
-          <TreeView.Item id="item-2-sub-task-2">
-            <TreeView.LeadingVisual label="Issue opened">
-              <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
-            </TreeView.LeadingVisual>
-            sub task 2
-          </TreeView.Item>
-        </TreeView.SubTree>
-      </TreeView.Item>
-      <TreeView.Item id="item-3">
-        <TreeView.LeadingAction>
-          <IconButton icon={GrabberIcon} aria-label="Reorder item 3" variant="invisible" />
-        </TreeView.LeadingAction>
-        <TreeView.LeadingVisual label="Issue opened">
-          <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
-        </TreeView.LeadingVisual>
-        Item 3
-      </TreeView.Item>
-    </TreeView>
-  )
-}
 
 export const MultilineItems: StoryFn = () => (
   <nav aria-label="Files changed">
