+++ packages/react/src/TreeView/TreeViewWithLeadingAction.stories.tsx
@@ -1,0 +1,71 @@
+import {GrabberIcon, IssueClosedIcon, IssueOpenedIcon} from '@primer/octicons-react'
+import type {Meta, StoryFn} from '@storybook/react'
+import React from 'react'
+import Box from '../Box'
+import {IconButton} from '../Button'
+import Octicon from '../Octicon'
+import {TreeView} from './TreeView'
+
+const meta: Meta = {
+  title: 'Private/Components/TreeViewWithLeadingAction',
+  component: TreeView,
+  decorators: [
+    Story => {
+      return (
+        <Box sx={{maxWidth: 400}}>
+          <Story />
+        </Box>
+      )
+    },
+  ],
+}
+
+export const LeadingAction: StoryFn = () => {
+  return (
+    <TreeView aria-label="Issues">
+      <TreeView.Item id="item-0">
+        <TreeView.LeadingAction>
+          <IconButton icon={GrabberIcon} aria-label="Reorder item 1" variant="invisible" />
+        </TreeView.LeadingAction>
+        <TreeView.LeadingVisual label="Issue closed">
+          <Octicon icon={IssueClosedIcon} sx={{color: 'done.fg'}} />
+        </TreeView.LeadingVisual>
+        Item 1
+      </TreeView.Item>
+      <TreeView.Item id="item-2">
+        <TreeView.LeadingAction>
+          <IconButton icon={GrabberIcon} aria-label="Reorder item 2" variant="invisible" />
+        </TreeView.LeadingAction>
+        <TreeView.LeadingVisual label="Issue opened">
+          <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
+        </TreeView.LeadingVisual>
+        Item 2
+        <TreeView.SubTree>
+          <TreeView.Item id="item-2-sub-task-1">
+            <TreeView.LeadingVisual label="Issue opened">
+              <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
+            </TreeView.LeadingVisual>
+            sub task 1
+          </TreeView.Item>
+          <TreeView.Item id="item-2-sub-task-2">
+            <TreeView.LeadingVisual label="Issue opened">
+              <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
+            </TreeView.LeadingVisual>
+            sub task 2
+          </TreeView.Item>
+        </TreeView.SubTree>
+      </TreeView.Item>
+      <TreeView.Item id="item-3">
+        <TreeView.LeadingAction>
+          <IconButton icon={GrabberIcon} aria-label="Reorder item 3" variant="invisible" />
+        </TreeView.LeadingAction>
+        <TreeView.LeadingVisual label="Issue opened">
+          <Octicon icon={IssueOpenedIcon} sx={{color: 'open.fg'}} />
+        </TreeView.LeadingVisual>
+        Item 3
+      </TreeView.Item>
+    </TreeView>
+  )
+}
+
+export default meta
